Fixed a bug that caused files to be processed out of order, cleaned up code
