+++ api/.env.example
@@ -1,0 +1,3 @@
+MONGODB_USERNAME=username
+MONGODB_PASSWORD=password
+PORT=8000
