+++ api/package.json
@@ -13,9 +13,12 @@
     "@mapbox/geojson-rewind": "^0.5.0",
     "axios": "^0.21.1",
     "cors": "^2.8.5",
+    "cron": "^1.8.2",
     "d3-contour": "^2.0.0",
+    "dotenv": "^8.2.0",
     "express": "^4.17.1",
     "geotiff": "^1.0.0-beta.16",
+    "mongodb": "^3.6.5",
     "nodemon": "^2.0.7",
     "simplify-geojson": "^1.0.4"
   }
